refactor : 인풋값 체크방법 변경
