+++ src/main/java/lotto/Application.java
@@ -4,11 +4,10 @@ import lotto.domain.*;
 import lotto.factory.LottoFactory;
 import lotto.generator.NumberGenerator;
 import lotto.generator.UserInputNumberGenerator;
-import lotto.utils.NumberUtils;
-import lotto.validator.Validator;
 import lotto.view.InputView;
 import lotto.view.OutputView;
 
+import java.util.Collections;
 import java.util.List;
 
 public class Application {
@@ -27,20 +26,21 @@ public class Application {
 
     private static ManualPurchase generateManualPurchase(Payment payment) {
         try {
-            int parsedManualCount = NumberUtils.parseNumber(InputView.getManualCount());
+            ManualPurchaseCount manualPurchaseCount = new ManualPurchaseCount(InputView.getManualCount(), payment);
+            List<String> manualLottoInput = getManualInput(manualPurchaseCount);
 
-            validateManualCount(parsedManualCount, payment);
-            List<String> manualLottoInput = InputView.getManualLottoNumbers(parsedManualCount);
-            return new ManualPurchase(parsedManualCount, manualLottoInput);
+            return new ManualPurchase(manualPurchaseCount, manualLottoInput);
         } catch (IllegalArgumentException exception) {
             OutputView.printErrorMessage(exception);
             return generateManualPurchase(payment);
         }
     }
 
-    private static void validateManualCount(int parsedManualCount, Payment payment) {
-        Validator.validateLessThanLottoCount(parsedManualCount, payment);
-        Validator.validatePositiveNumber(parsedManualCount);
+    private static List<String> getManualInput(ManualPurchaseCount manualPurchaseCount) {
+        if (manualPurchaseCount.getPurchasedCount() > 0) {
+            return InputView.getManualLottoNumbers(manualPurchaseCount);
+        }
+        return Collections.emptyList();
     }
 
     private static Payment generatePayment() {
@@ -57,6 +57,7 @@ public class Application {
             NumberGenerator numberGenerator = new UserInputNumberGenerator();
             String winningLottoInput = InputView.getWinningLottoNumber();
             String bonusNumberInput = InputView.getBonusNumber();
+
             return new WinningLottoTicket(numberGenerator.generateNumbers(winningLottoInput), new LottoNumber(bonusNumberInput));
         } catch (IllegalArgumentException exception) {
             OutputView.printErrorMessage(exception);
